https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved mailboxes services to esn.inbox.libs
