+++ src/esn.inbox.libs/app/services/mailboxes/special-mailboxes.run.js
@@ -1,0 +1,8 @@
+'use strict';
+
+require('./special-mailboxes.constants.js');
+
+angular.module('esn.inbox.libs')
+  .run(function(inboxSpecialMailboxes, INBOX_ALL_MAIL_MAILBOX) {
+    inboxSpecialMailboxes.add(INBOX_ALL_MAIL_MAILBOX);
+  });
