+++ src/linagora.esn.unifiedinbox/app/directives/lists.js
@@ -4,7 +4,7 @@ require('../../../esn.inbox.libs/app/services/config/config.js');
 require('../services/new-composer/new-composer.js');
 require('../services/jmap-item/jmap-item-service.js');
 require('../services.js');
-require('../services/mailboxes/mailboxes-service.js');
+require('../../../esn.inbox.libs/app/services/mailboxes/mailboxes-service.js');
 
 (function (angular) {
   'use strict';
