+++ src/linagora.esn.unifiedinbox/app/directives/main.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 require('../services/new-composer/new-composer.js');
 require('../../../esn.inbox.libs/app/services/custom-role-mailbox/custom-role-mailbox.service.js');
-require('../services/mailboxes/mailboxes-service.js');
+require('../../../esn.inbox.libs/app/services/mailboxes/mailboxes-service.js');
 require('../services/jmap-item/jmap-item-service.js');
 require('../../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('../services/shortcuts/shortcuts.constants.js');
