+++ src/linagora.esn.unifiedinbox/app/services/filtering/filtering-service.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-require('../mailboxes/mailboxes-service.js');
+require('../../../../esn.inbox.libs/app/services/mailboxes/mailboxes-service.js');
 require('../filtering/filters.js');
 
 (function(angular) {
