--- src/linagora.esn.unifiedinbox/app/services/mailboxes/shared-mailboxes.constants.js
@@ -1,15 +1,0 @@
-(function(angular) {
-  'use strict';
-
-  angular.module('linagora.esn.unifiedinbox')
-
-    .constant('INBOX_HIDDEN_SHAREDMAILBOXES_CONFIG_KEY', 'hiddenSharedMailboxes')
-    .constant('INBOX_ROLE_NAMESPACE_TYPES', {
-      shared: 'delegated',
-      owned: 'personal'
-    })
-    .constant('INBOX_MAILBOXES_NON_SHAREABLE', ['drafts', 'outbox', 'sent', 'trash'])
-    .constant('INBOX_DEFAULT_FOLDERS_SHARING_CONFIG', false)
-    .constant('INBOX_FOLDERS_SHARING_CONFIG_KEY', 'features.foldersSharing');
-
-})(angular);
